support OID related functionaltiy in CAS and DB_SERVER
