+++ src/method/method_callback.hpp
@@ -37,33 +37,6 @@
 
 namespace cubmethod
 {
-  /* OID_CMD are ported from cas_cci.h */
-  enum OID_CMD
-  {
-    OID_CMD_FIRST = 1,
-
-    OID_DROP = 1,
-    OID_IS_INSTANCE = 2,
-    OID_LOCK_READ = 3,
-    OID_LOCK_WRITE = 4,
-    OID_CLASS_NAME = 5,
-
-    OID_CMD_LAST = OID_CLASS_NAME
-  };
-
-  enum COLLECTION_CMD
-  {
-    COL_CMD_FIRST = 1,
-    COL_GET = 1,
-    COL_SIZE = 2,
-    COL_SET_DROP = 3,
-    COL_SET_ADD = 4,
-    COL_SEQ_DROP = 5,
-    COL_SEQ_INSERT = 6,
-    COL_SEQ_PUT = 7,
-    COL_CMD_LAST = COL_SEQ_PUT
-  };
-
   struct lob_handle
   {
     int db_type;
@@ -88,20 +61,12 @@ namespace cubmethod
       int oid_get (packing_unpacker &unpacker);
       int oid_put (packing_unpacker &unpacker);
       int oid_cmd (packing_unpacker &unpacker);
-
       int collection_cmd (packing_unpacker &unpacker);
 
       void set_server_info (int rc, char *host);
       void free_query_handle_all ();
 
     protected:
-      int col_get (DB_COLLECTION *col, DB_TYPE col_type, DB_TYPE ele_type, DB_DOMAIN *ele_domain);
-      int col_size (DB_COLLECTION *col);
-      int col_set_drop (DB_COLLECTION *col, DB_VALUE *ele_val);
-      int col_set_add (DB_COLLECTION *col, DB_VALUE *ele_val);
-      int col_seq_drop (DB_COLLECTION *col, int seq_index);
-      int col_seq_insert (DB_COLLECTION *col, int seq_index, DB_VALUE *ele_val);
-      int col_seq_put (DB_COLLECTION *col, int seq_index, DB_VALUE *ele_val);
 
 // TODO: implement remaining functions on another issues
 #if 0
@@ -111,8 +76,6 @@ namespace cubmethod
 #endif
 
     private:
-      int check_object (DB_OBJECT *obj);
-
       /* ported from cas_handle */
       int new_query_handler ();
       query_handler *find_query_handler (int id);
